feat: splitting LLMClient as a separated module
